IQuantity with QuantityValue as Value type (PoC) - QuantityValue implements most of the INumber interface members / operators (operations  preserving decimal when possible, failing back to double, clamped to the real values- infinities excluded) - changed the type of IQuantity.Value/As to QuantityValue (explicit cast required) - added back the IEquatable interface: behavior depending on the runtime type of the quantity values (strict equality comparison for double/double) - added back the fix for the non-reflexive CompareTo issue (and the corresponding tests) - added a check preventing the possible stack overflow exception when calling ToUnit() with default(TUnit) (or any other invalid Enum value) - DataContracts changed to reflect the possible use of either double or decimal as ValueType (the Json converters were not changed- but probably should be) - removed the obsolete Json converter tests

+++ UnitsNet.Serialization.JsonNet/UnitsNetBaseJsonConverter.cs
@@ -186,13 +186,14 @@ namespace UnitsNet.Serialization.JsonNet
                 return new ExtendedValueUnit
                 {
                     Unit = $"{quantity.QuantityInfo.UnitType.Name}.{quantity.Unit}",
-                    Value = quantity.Value,
+                    // The type of "Value" is still double
+                    Value = (double)quantity.Value,
                     ValueString = d.Value.ToString(CultureInfo.InvariantCulture),
                     ValueType = "decimal"
                 };
             }
 
-            return new ValueUnit {Value = quantity.Value, Unit = $"{quantity.QuantityInfo.UnitType.Name}.{quantity.Unit}"};
+            return new ValueUnit {Value = (double)quantity.Value, Unit = $"{quantity.QuantityInfo.UnitType.Name}.{quantity.Unit}"};
         }
 
         /// <summary>
